+++ grafana_wtf/report.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 # (c) 2019 Andreas Motl <andreas@hiveeyes.org>
 # License: GNU Affero General Public License, Version 3
-import logging
 import colored
+import logging
+from six import StringIO
 from urllib.parse import urljoin
+from collections import OrderedDict
 from grafana_wtf.util import prettify_json
 
 log = logging.getLogger(__name__)
@@ -33,22 +35,32 @@ class WtfReport:
         for item in items:
             #print('item:', item)
 
+            if self.verbose:
+                print()
+                print(prettify_json(item.data))
+
             # Output match title / entity name.
             name = self.get_item_name(item)
             print(_v(name))
             print('-' * len(name))
 
-            print(_vlow(url_callback(item)))
+            # Output URL
+            url = url_callback(item)
+
+            # Output baseline bibliographic data.
+            print()
+            bibdata_output = self.get_bibdata(item, URL=_vlow(url))
+            if bibdata_output:
+                print(bibdata_output)
+
+            # Output findings.
             if 'matches' in item.meta:
                 #print(' ', self.format_where(item))
-                print()
                 for match in item.meta.matches:
                     print('- {path}: {value}'.format(path=_k(match.full_path), value=_v(match.value)))
-            if self.verbose:
-                print()
-                print(prettify_json(item.data))
-            else:
-                print()
+
+            print()
+            print()
 
     def get_item_name(self, item):
         if 'name' in item.data:
@@ -57,6 +69,26 @@ class WtfReport:
             return item.data.meta.slug
         else:
             return 'unknown'
+
+    def get_bibdata(self, item, **kwargs):
+
+        # Sanity checks.
+        if 'dashboard' not in item.data:
+            return
+
+        bibdata = OrderedDict()
+        bibdata['Title'] = item.data.dashboard.title
+        bibdata['Folder'] = item.data.meta.folderTitle
+        bibdata['UID'] = item.data.dashboard.uid
+        bibdata['Created'] = f'at {item.data.meta.created} by {item.data.meta.createdBy} '
+        bibdata['Updated'] = f'at {item.data.meta.updated} by {item.data.meta.updatedBy}'
+        bibdata.update(kwargs)
+        output = StringIO()
+        for key, value in bibdata.items():
+            entry = f' {key:>7} {value}\n'
+            output.write(entry)
+        output.seek(0)
+        return output.read()
 
     def format_where(self, item):
         keys = item.meta.where
@@ -72,6 +104,16 @@ class WtfReport:
     def compute_url_dashboard(self, dashboard):
         return urljoin(self.grafana_url, dashboard.data.meta.url)
 
+    def experimental(self):
+        # print(match)
+        # print(dir(match.context))
+        # print(match.context)
+        # pprint(match.context.value)
+        # idp = parse('id')
+        # if match.context.parse('id'):
+        #    print('  id:', match.context.id)
+        # pprint(match)
+        pass
 
 section_style = colored.fg("cyan") + colored.attr("bold")
 key_style = colored.fg("blue") + colored.attr("bold")